Revert "リソース読み込み関連オブジェクトをひとつのファイルにまとめた"
This reverts commit 3c0333ef185c50e99fd7a520ee58624dbe3aebd6.
