+++ src/js/game/dom-scenes/loading/index.js
@@ -1,6 +1,6 @@
 // @flow
 
-import type {LoadingActions, LoadingProgress} from "../../../resource/loading";
+import type {LoadingActions, LoadingProgress} from "../../../resource/loading-actions";
 import {LoadingPresentation} from "./presentation";
 import type {DOMScene} from "../dom-scene";
 import type {Stream, Unsubscriber} from "../../../stream/core";
